Split Ftype into Ftype and FEtype (#3341)
* fix include directories not find bug (#3323)



* reduce grpc download file size from 380M to 130M  (#3326)

* reduce grpc download file size from 380M to 130M


* fix bug



* update clang-tidy config (#3314)

* update clang-tidy config


* update .clang-tidy


* update clang-tidy config


* update clang-tidy HeaderFilterRegex


* enable fetype


* rename field_element_method


* add FETYPE_TYPE


* Add WebServer unittest (#3321)

* Add web server interface


* Add unittest/server


* Add web server ut


* modify db schema and remove conversion


* weak fetype


* enforce strong fetype


* lint


* format code


* [skip ci] fix

+++ core/src/server/web_impl/handler/WebRequestHandler.cpp
@@ -821,7 +821,7 @@ WebRequestHandler::Search(const std::string& collection_name, const nlohmann::js
                 for (const auto& field : field_mappings) {
                     auto field_name = field.first->GetName();
                     auto field_data = result->data_chunk_->fixed_fields_;
-                    switch ((int64_t)field.first->GetFtype()) {
+                    switch (field.first->GetFtype()) {
                         case engine::DataType::INT32: {
                             int32_t int32_value;
                             int64_t offset = (i * step + j) * sizeof(int32_t);
